Cross-platform utils moved to it's own directory. Tests add .NET 4.7.1 target
