+++ RedisCore/Internal/Protocol/RedisByteString.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 #if NETSTANDARD2_0
-using RedisCore.Utils;
+using RedisCore.CrossPlatform;
 #endif
 
 namespace RedisCore.Internal.Protocol
